backend/requirements.txt
flask
flask-cors
msgspec
pandas
geopandas
shapely
networkx
gunicorn
pyodbc
pyyaml